Revert "Support more than 5 mouse buttons" (FS#1082)
This reverts commit bd8158495e059baa0e8ba525b813254451447f4a.

The idea was to track the current list of pressed and depressed mouse buttons,
because we get button events for more than 5 buttons, but can only query the
state of the first 5 buttons.

However, there are cases where we see button presses, but won't see the
corresponding release event. This is quite bad.

Signed-off-by: Uli Schlachter <psychon@znc.in>