Fix focus tracking on tag viewing
The original reason for this patch was the following:
When focusing a client and hiding that by unviewing all of its tags
and then trying to set the urgent hint on that client got wrongly
ignored.

It also (slightly) changes behaviour when viewing tags: If a new client
is made visible by viewing a tag, and that client comes before the
currently focused one in the focus list, it is focused instead of the
current one.

Signed-off-by: Gregor Best <gbe@ring0.de>
Signed-off-by: Julien Danjou <julien@danjou.info>