Switch from libXcursor to libxcb-cursor
Thanks to Michael Stapelberg, there is now a xcb-only port of libXcursor which
does everything we need. This patch switches awesome over to that new library.

Since the only reason for using XOpenDisplay() instead of xcb_connect() was so
that we can use libXcursor, we can get back to that older state again. This
means that this effectively reverts the following commits:

531f8b415c65e9754a907e72f984ab0f6cbfc3a9 "Added initial support for Xlib cursor themes"
77243cd09aa862da10d18f6d549ef10d3c775d2b "Add x11-xcb to the pkg-config checks"
779d43fc462411d84adf757a178e938c6ba9d3b8 "Don't let Xlib own the event queue"
03759b48478eb716af8b62e01ef3b5140c582f58 "Fix keyboard layouts"

Signed-off-by: Uli Schlachter <psychon@znc.in>